Weapon reuse delay and hit timer corrections: - Bow SA quick recovery was used when calculating the gauge, but it was not used in calculating the start of the next attack (resulting in extra gap between attacks). - When a hit was made, the AI timed event calculating the damage and sending client packets hitmissed/youdiddamage/... was neglecting the actual attack speed. The time delays were hard coded. Now the AI event is run at the actual time of the hit.
A matter of opinion: If you feel that the server packet about the hit damage comes too late,
we could run the event like 100 ms before the actual hit. Now there's some gap between hit
animation and info packet. With the old hard coded values it varies... with low attack speeds
the info was sent before the hit, with high attack speeds it was after.